Backend/requirements.txt
# For MongoDB and password hashing
motor
passlib[bcrypt]
fastapi
uvicorn
pydantic
python-multipart
torch
torchvision
Pillow
opencv-python-headless
numpy
timm
geopy
python-dotenv
requests
beautifulsoup4
transformers==4.39.3
sentence-transformers==2.6.1
email-validator
requests-html
lxml[html_clean]